requirements.txt
aiogram==3.14.0
playwright==1.55.0
python-dotenv==1.0.1